pyproject.toml
[project]
name = "workflow-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pygithub>=2.6.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
]
